TL-122 - “Help” button is not working correctly
The reference to the dialog was disapearing cause jqueryui moves
the div element that represents the dialog to a different place in
the DOM and the code that open the dialog was using relative location
to find the dialog element.

I change that to use the HTML ID instead so no matter where is move
in the DOM it would always have a direct access to it.